requirements.txt
streamlit
google-cloud-bigquery
pandas
python-dateutil
protobuf